loading cracCreationParameters in each runner
