--- cse-lib/network-processing/src/test/resources/com/farao_community/farao/cse/network_processing/cseCracCreationParameters.json
@@ -1,14 +1,0 @@
-{
-  "crac-factory": "CracImplFactory",
-  "ra-usage-limits-per-instant": [
-    {
-      "instant": "curative",
-      "max-ra-per-tso": {
-        "IT": 6,
-        "FR": 5,
-        "CH": 1,
-        "AT": 3,
-        "SI": 3
-      }
-    }]
-}
